Add wltv and wapr calculation
