+++ src/pages/tokenLending/LoansTokenPage/TokenLoansContent/components/ExpandedCardContent/helpers.ts
@@ -1,16 +1,13 @@
-import { web3 } from 'fbonds-core'
-import { calcBorrowerTokenAPR } from 'fbonds-core/lib/fbond-protocol/helpers'
-import { map, sumBy } from 'lodash'
+import { sumBy } from 'lodash'
 
 import { TokenLoan } from '@banx/api/tokens'
 import {
   caclulateBorrowTokenLoanValue,
   calcTokenWeeklyFeeWithRepayFee,
-  calcWeightedAverage,
   isTokenLoanRepaymentCallActive,
 } from '@banx/utils'
 
-import { calcTokenTotalValueToPay } from '../../helpers'
+import { calcTokenTotalValueToPay, calculateWeightedApr } from '../../helpers'
 
 export const getPayInterestActionText = (loans: TokenLoan[]) => {
   if (loans.length === 0) return 'Pay'
@@ -22,17 +19,6 @@ export const getPayInterestActionText = (loans: TokenLoan[]) => {
   if (noActiveRepaymentCalls) return 'Pay interest'
 
   return 'Pay'
-}
-
-const calculateWeightedApr = (loans: TokenLoan[]) => {
-  const totalAprValues = map(loans, (loan) => {
-    const marketPubkey = new web3.PublicKey(loan.fraktBond.hadoMarket)
-    return calcBorrowerTokenAPR(loan.bondTradeTransaction.amountOfBonds, marketPubkey) / 100
-  })
-
-  const totalRepayValues = map(loans, (loan) => caclulateBorrowTokenLoanValue(loan).toNumber())
-
-  return calcWeightedAverage(totalAprValues, totalRepayValues)
 }
 
 export const calculateLoansStats = (loans: TokenLoan[]) => {
